Add works to batch UI
This changes the files view to include `PreIngestWorks`.

The files are grouped by their `PreIngestWork`, which includes
the `deduplication_key` and `date`. This also includes the `title`,
which is looked up in solr via the `deduplication_key`. If this
isn't present for a work it means that the work has not been indexed
yet so a message saying that is displayed instead of the title.

Additional attributes from solr could be added in the future.

Connected to https://github.com/curationexperts/in-house/issues/426