Change count language on preview screen
The format of the CSV files has changed so that
a single row is not a work. The count available
on the preview screen is a count of rows. This
changes the language to reflect that the number
is a count of rows and not records/works.

Connected to https://github.com/curationexperts/in-house/issues/43