Basic file status
This allows a PreIngestFile to look for a FileSet in
solr based on the sha1 of the file. Right now that's
the most unique thing we can use to identify whether the
file has been indexed in solr. If the hash is found it displays
a green checkmark in the files table.

If the hash isn't found it displays a question mark icon.

Connected to https://github.com/curationexperts/in-house/issues/423