Basic file status
This adds a PreIngestWork controller that allows
you to get the thumbnail URLs for a work based on
the deduplication_key.

This is used by some JS that checks the thumbnail URLs
by only requesting headers from the URL. If the work's
files have been characterized the requests will return
200. That's used to change change the icon from a question
mark to a green check in the 'Works and Files' page.

This kind of status checking can't be used for individual files
right now because there isn't any unique information being
saved on the `FileSet` that would be present when the `PreIngestFile`
is created.

Connected to https://github.com/curationexperts/in-house/issues/423