More detailed file status
This adds a status field to `PreIngestFile`s and
`PreIngestWorks` that defaults to 'preingest'.

The `ContentDepositEventJob` and `FileSetAttachedEventJob` classes
are preprended with logic that updates their status to 'attached'.

The logic requires that `deduplication_key` is present in the
CSV so that field is added to the default mapper and validator.

There is a helper which outputs a success icon (a green checkmark)
if a `PreIngestFile` or `PreIngestWork` has an 'attached' status.