Fix endless dependency calculation
Zizia currently has a bug where new installations do not work because
the dependency calculation hangs forever. As far as I can tell this is
somehow caused by the dependency on web-console. When I remove it,
dependencies start getting calculated without problem. Since web-console
isn't mission critical to zizia, let's just remove it for now and see if
that fixes the problem.